View 분리
